requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pdfplumber
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-multipart
google-genai
pandas
